+++ extern/CSparse/Tcov/cs_pinv.c
@@ -1,0 +1,1 @@
+../Source/cs_pinv.c